# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Official Otterly API documentation.",
        "title": "Otterly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/users": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get all users data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-array_github_com_otterly-id_otterly_backend_internal_api_models_UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add new user data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "Create user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get user data based on provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Remove user data based on provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponseWithoutData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Edit user data based on provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_otterly-id_otterly_backend_internal_api_models.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 8
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 20
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "USER",
                        "OWNER"
                    ]
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.CreateUserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UserRole"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.FailureResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-array_github_com_otterly-id_otterly_backend_internal_api_models_UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UserResponse"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_CreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.CreateUserResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_UpdateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UpdateUserResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponse-github_com_otterly-id_otterly_backend_internal_api_models_UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UserResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.SuccessResponseWithoutData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UserRole"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_otterly-id_otterly_backend_internal_api_models.UserRole"
                }
            }
        },
        "github_com_otterly-id_otterly_backend_internal_api_models.UserRole": {
            "type": "string",
            "enum": [
                "ADMIN",
                "USER",
                "OWNER"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser",
                "RoleOwner"
            ]
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT token stored in httpOnly cookie for authentication",
            "type": "apiKey",
            "name": "otterly_token",
            "in": "cookie"
        }
    },
...
}